requirements.txt
annotated-types==0.7.0
anyio==4.8.0
Authlib==1.4.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.1
distro==1.9.0
fastapi==0.115.8
gitignore_parser==0.1.11
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
MarkupSafe==3.0.2
numpy==1.26.4
openai==1.63.0
packaging==24.2
pandas==2.2.3
plotly==5.24.1
pyarrow==15.0.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
tenacity==9.0.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
watchdog==3.0.0
websockets==12.0
writer==0.8.2
writer-sdk==1.6.1
